Bug fixes
Changed length of signature for Windows executables to 2 bytes, the 3rd and onward bytes are not guaranteed (thus did not successfully scan msblast.exe). Also changed the position of the 'scanning...' message to before it attempts to load the file to memory. This makes output a bit friendlier since you are no longer waiting for it to map before you are told it is even trying to scan it.